framework/qatron-python/pyproject.toml
[tool.poetry]
name = "qatron-python"
version = "0.1.0"
description = "QAtron Python automation framework"
authors = ["QAtron Team"]

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.4.0"
pytest-bdd = "^7.0.0"
pytest-cov = "^4.1.0"
pytest-html = "^4.1.0"
pytest-xdist = "^3.5.0"
allure-pytest = "^2.13.2"
selenium = "^4.15.0"
requests = "^2.31.0"
pyyaml = "^6.0.1"
coverage = "^7.3.4"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
ruff = "^0.1.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "unit: Unit tests",
    "contract: Contract tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "suite: Test suite marker",
    "flaky_candidate: Potentially flaky test",
]

[tool.coverage.run]
source = ["."]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
